upd test_nightly_vlm_models_perf.py
