+++ python/sglang/test/test_utils.py
@@ -1467,3 +1467,29 @@ def dump_bench_raw_result(
 def _ensure_remove_suffix(text: str, suffix: str):
     assert text.endswith(suffix)
     return text.removesuffix(suffix)
+
+
+# results are from `bench_one_batch_server.py`
+def generate_markdown_report(model, results, input_len, output_len):
+    summary = f"### {model}\n"
+    summary += f"Input lens: {input_len}. Output lens: {output_len}.\n"
+    summary += "| batch size | latency (s) | input throughput (tok/s)  | output throughput (tok/s) | acc length | ITL (ms) | input cost ($/1M) | output cost ($/1M) | profile |\n"
+    summary += "| ---------- | ----------- | ------------------------- | ------------------------- | ---------- | -------- | ----------------- | ------------------ |-------------|\n"
+
+    for result in results:
+        # Extract the metrics row that bench_one_batch_server prints (without the profile column)
+        m = re.search(
+            r"\|\s*([\d\.]+)\s*\|\s*([\d\.]+)\s*\|\s*([\d\.]+)\s*\|\s*([\d\.]+)\s*\|\s*(?:n/a|[\d\.]+)\s*\|\s*([\d\.]+)\s*\|\s*([\d\.]+)\s*\|\s*([\d\.]+)\s*\|",
+            result["output"],
+        )
+        if m:
+            # Reconstruct the row and append a placeholder artifact link for profile
+            parts = [part.strip() for part in m.group(0).split("|") if part.strip()]
+            PROFILE_URL_PLACEHOLDER = "PROFILE_URL_PLACEHOLDER"
+            row = f"| {' | '.join(parts)} | [Profile]({PROFILE_URL_PLACEHOLDER}) |\n"
+            summary += row
+    return summary
+
+
+def parse_models(model_string: str):
+    return [model.strip() for model in model_string.split(",") if model.strip()]
